src/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
bcrypt==4.0.1
redis==5.0.0
cryptography==41.0.7